requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
contextvars==2.4
dataclasses==0.6
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.11
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
entrypoints==0.3
idna==3.3
immutables==0.16
importlib-metadata==4.8.3
iniconfig==1.1.1
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==3.0.3
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.1.1
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
MarkupSafe==2.0.1
mistune==0.8.4
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.7
packaging==21.3
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2==2.7.7
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
pytest-django==4.5.2
pytest-sugar==0.9.4
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
zipp==3.6.0

pip~=21.1.2
wheel~=0.36.2
setuptools~=57.0.0
django-simple-history~=2.8.0
future~=0.18.2
zappa~=0.54.1
django-environ==0.8.1